修改时间参数类型为long

+++ data/src/main/java/io/basc/framework/data/TemporaryCounter.java
@@ -2,7 +2,7 @@ package io.basc.framework.data;
 
 public interface TemporaryCounter extends Counter{
 	
-	long incr(String key, long delta, long initialValue, int exp);
+	long incr(String key, long delta, long initialValue, long exp);
 
-	long decr(String key, long delta, long initialValue, int exp);
+	long decr(String key, long delta, long initialValue, long exp);
 }
